Reworded "Why" section
Rearranged the "Why" section to incorporate feedback from @lewiz and @exorus, now that I understand the purpose of the extension better.

Kept Postfix in as an example so it would come up on a keyword search. I recommend working in other keywords people may be searching for to increase visibility.