Save Attachment Enhancement.
Instead of giving an attachment a unique filename and extension 'tmp'
when the original fileName already exists, the attachment is now saved
as the original fileName with a integer suffix and original extension.

When the suffix reaches the number 1000 and the fileName still already
exist, the fileName will be random and unique, but the extension still
remains the original one.