Fix encoding from Windows-based software products
There is an issue concerning the simplified chinese encoding.
As explained in this article (https://en.wikipedia.org/wiki/Code_page_936_(Microsoft_Windows)), there is a confusion between GB2312, CP936 and GBK with Windows-based software products.

In my case, my application fetches an email from outlook.office365 with the header "Content-Type: text/html; charset="gb2312"". But decoding with "gb2312" does not work, and I have to decode with gbk. That works perfectly.

If you have a better solution that the one I propose, feel free to improve it :)